Support creating and removing a network before/after the feature
The idea behind this is to attach test containers to this network, but
they currently do not support attaching their containers to a network
different than bridge.

We have opened an issue in the testcontainers-go repo (https://github.com/testcontainers/testcontainers-go/issues/73)
and have a working branch adding that support (https://github.com/mdelapenya/testcontainers-go/tree/support-for-networking). We are waiting for feedback before sending it